+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -1707,14 +1707,12 @@ class CreatePromptVersionsTypedDict(TypedDict):
     r"""Prompt version model returned from the API"""
 
     id: str
-    created_by_id: str
-    display_name: str
-    r"""The prompt’s name, meant to be displayable in the UI."""
-    updated_by_id: str
     prompt_config: CreatePromptPromptsResponsePromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: CreatePromptPromptsResponseMetadataTypedDict
     timestamp: str
+    created_by_id: NotRequired[str]
+    updated_by_id: NotRequired[str]
     description: NotRequired[Nullable[str]]
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
 
@@ -1724,13 +1722,6 @@ class CreatePromptVersions(BaseModel):
 
     id: Annotated[str, pydantic.Field(alias="_id")]
 
-    created_by_id: str
-
-    display_name: str
-    r"""The prompt’s name, meant to be displayable in the UI."""
-
-    updated_by_id: str
-
     prompt_config: CreatePromptPromptsResponsePromptConfig
     r"""A list of messages compatible with the openAI schema"""
 
@@ -1738,12 +1729,16 @@ class CreatePromptVersions(BaseModel):
 
     timestamp: str
 
+    created_by_id: Optional[str] = None
+
+    updated_by_id: Optional[str] = None
+
     description: OptionalNullable[str] = UNSET
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = ["description"]
+        optional_fields = ["created_by_id", "updated_by_id", "description"]
         nullable_fields = ["description"]
         null_default_fields = []
 
@@ -1778,15 +1773,15 @@ class CreatePromptResponseBodyTypedDict(TypedDict):
     id: str
     owner: CreatePromptOwnerTypedDict
     domain_id: str
-    created_by_id: str
     display_name: str
     r"""The prompt’s name, meant to be displayable in the UI."""
-    updated_by_id: str
     prompt_config: CreatePromptPromptsPromptConfigTypedDict
     r"""A list of messages compatible with the openAI schema"""
     metadata: CreatePromptPromptsMetadataTypedDict
     type: CreatePromptPromptsType
     versions: List[CreatePromptVersionsTypedDict]
+    created_by_id: NotRequired[str]
+    updated_by_id: NotRequired[str]
     description: NotRequired[Nullable[str]]
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
     created: NotRequired[datetime]
@@ -1804,12 +1799,8 @@ class CreatePromptResponseBody(BaseModel):
 
     domain_id: str
 
-    created_by_id: str
-
     display_name: str
     r"""The prompt’s name, meant to be displayable in the UI."""
-
-    updated_by_id: str
 
     prompt_config: CreatePromptPromptsPromptConfig
     r"""A list of messages compatible with the openAI schema"""
@@ -1820,18 +1811,28 @@ class CreatePromptResponseBody(BaseModel):
 
     versions: List[CreatePromptVersions]
 
+    created_by_id: Optional[str] = None
+
+    updated_by_id: Optional[str] = None
+
     description: OptionalNullable[str] = UNSET
     r"""The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose"""
 
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T14:13:16.580Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T14:20:21.534Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
     def serialize_model(self, handler):
-        optional_fields = ["description", "created", "updated"]
+        optional_fields = [
+            "created_by_id",
+            "updated_by_id",
+            "description",
+            "created",
+            "updated",
+        ]
         nullable_fields = ["description"]
         null_default_fields = []