+++ packages/orq-rc/RELEASES.md
@@ -289,3 +289,13 @@ Based on:
 - [python v3.1.0-rc.51] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.51] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.51 - packages/orq-rc
+
+## 2025-02-10 14:21:51
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.487.1 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.52] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.52] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.52 - packages/orq-rc